Fix form URL error
to_param must be implemented to explicitly delegate.  Why?  Because
Rails automatically creates to_param for all objects, forwarding to
their to_s method.  For the DataStructure stuff, this causes to_param
to be called on the parent class - delegation only happens when the
parent does *not* have the given method.

Not sure what other fun issues we'll encounter, but this could be the
beginning of some major Rails magic pain.