Use draper decoration instead of fully custom
Splits up decorator class so it subclasses Draper and then includes
the DataStructure::Container module.  This separation should mean we
can use DataStructure::Container directly on a model if desired, but
also removes some of the delegation magic because Draper handles it
for us.

Note that by using draper, naming conventions can get pretty weird.
It's worth considering the "decorates ..." method as is now in the
example and tests, as that avoids weird errors attempting to look up
names of classes which don't exist.